update student courses to publish when student registers for courses.
In the past, scores are only published when the academics admin publishes score. The downside to this is if student registers after scores have been published, student wouldn't see her scores. So it is important that as soon as student registers, her courses should default to whatever publish status is available for that course.